backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
psycopg[binary]>=3.2
python-dotenv>=1.0
